\name{getSupportedCancerTypes}
\alias{getSupportedCancerTypes}
\title{Get supported cancer types}
\description{
Retrieves the names of cancer types supported by the \code{\link{tpacForCancer}} function.
}
\usage{
    getSupportedCancerTypes()
}
\value{
Vector containing the names of supported cancer types.
}
\seealso{\code{\link{tpacForCancer}}}

\keyword{file}
\examples{
  getSupportedCancerTypes()
}