\name{TPAC-package}
\alias{TPAC-package}
\docType{package}
\title{
Tissue-adjusted pathway analysis of cancer (TPAC)
}
\description{
Implementation of TPAC, a method for single sample gene set testing of cancer transcriptomics data that leverages normal tissue-specificity.
}
\details{
\tabular{ll}{
Package: \tab TPAC\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2023\cr
License: \tab GPL-2 \cr
}
}
\author{
H. Robert Frost
}
\references{
    \itemize{
    \item Frost, H. R. (2022). Tissue-adjusted pathway analysis of cancer (TPAC). biorXiv e-prints. doi: https://doi.org/10.1101/2022.03.17.484779 
    }
}
\note{
This work was supported by the National Institutes of Health grants R35GM146586, R21CA253408, P20GM130454 and P30CA023108.
}
\keyword{ package }

